+++ server/catgenome/src/main/java/com/epam/catgenome/manager/wig/reader/BedGraphFeature.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2016 EPAM Systems
+ * Copyright (c) 2017 EPAM Systems
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,6 +26,9 @@ package com.epam.catgenome.manager.wig.reader;
 
 import htsjdk.tribble.Feature;
 
+/**
+ * Represent a record from BedGraph files
+ * */
 public class BedGraphFeature implements Feature {
 
     /**
